+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* **Feature**: `@subscribe()` can now also take promises for event targets and promise-returning factories as its first argument
 * **Bugfix**: Ensure that *all* methods decorated with `@init()` in *all* classes in an inheritance chain only fire when the *last* enhanced constructor finishes.
 * **Other**: Bump depdencies, play some code golf, tweak error messages